agent-api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
openai-agents==0.2.1
python-dotenv==1.0.0
pydantic==2.10.3
